[Frontend] use stol

+++ nnvm/Frontend/IRGenerator.cpp
@@ -1513,9 +1513,10 @@ Any IRGenerator::visitExp(SysYParser::ExpContext *ctx) {
     }
 
     if (auto *intConst = number->INTEGER_CONST()) {
-      return Symbol{createConstInt(std::stoi(intConst->getText(), 0,
-                                             getRadixOf(intConst->getText()))),
-                    SymbolType::getIntTy()};
+      return Symbol{
+          createConstInt((int)(std::stol(intConst->getText(), 0,
+                                         getRadixOf(intConst->getText())))),
+          SymbolType::getIntTy()};
     }
 
     nnvm_unreachable("No such literal number")
